Fixed a bunch of issues and optimized the code.
Upon being bankrupt, a player was not being properly removed, I implemented Java's Iterator to remove a player successfully.

I updated the code to show player the price of a property before they purchase.

Previously, when a player was being taxed, the owner of the property was not receiving the taxed amount, this is now fixed.

I removed 'buy' from CommandWords.java since the game already prompts user to purchase when necessary.